requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-mgmt-core==1.4.0
azure-mgmt-network==23.0.0
azure-mgmt-resource==23.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
pywin32==306
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
